+++ test/suite/helpers/common.js
@@ -85,25 +85,6 @@ testStaticAndInstance = function (subject, args, expected, message) {
   }
 }
 
-testGetSparseArray = function(index) {
-  var arr = [];
-  for (var i = 1; i < arguments.length; i++) {
-    arr[index + i - 1] = arguments[i];
-  }
-  return arr;
-}
-
-// Using the [] constructor with undefined values in <= IE8 will effectively
-// produce a sparse array (no property actually gets set) the only workaround
-// is to push undefined directly onto the array.
-testGetArrayWithUndefined = function() {
-  var arr = [];
-  for (var i = 0; i < arguments.length; i++) {
-    arr.push(arguments[i]);
-  }
-  return arr;
-}
-
 // iOS 8 supports Set but not the iterable constructor syntax (it simply adds
 // the array as a set member instead). So this helper method helps creates sets
 // from arguments.